Added make rule to compile .m into .o files.
Once the .m routines are installed, we compile them all into the VistA instance
/o directory, using a make rule. This new rule invokes the mumps compiler with
the environment variable gtm_destdir set to remove the prefix of the local
debian build directory. In this way, the resulting .o files can be relocated to
their final destination in the destdir/o instance directory, and yet the $TEXT
function will be able to locate the source code of the routines.